Only `ssh-keyscan` into `known_hosts` if necessary
Before this PR, we would `ssh-keyscan` unconditionally, which slows us down a bit, and also creates enormous `known_hosts` files over time, if your runners are not ephemeral.  It's much better to first check (via `ssh-keygen -F`) if a host is not found in your `known_hosts` file, and only then update it.